+++ lib/simple_3d_renderer.dart
@@ -1,6 +1,7 @@
 export 'src/sp3d_camera.dart';
 export 'src/sp3d_orthographic_camera.dart';
 export 'src/sp3d_camera_zoom_controller.dart';
+export 'src/sp3d_camera_rotation_controller.dart';
 export 'src/sp3d_faceobj.dart';
 export 'src/sp3d_gesture_detector.dart';
 export 'src/sp3d_light.dart';